% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmerpenguins-package.R
\docType{package}
\name{palmerpenguins-package}
\alias{palmerpenguins}
\alias{palmerpenguins-package}
\title{palmerpenguins: Palmer Archipelago (Antarctica) Penguin Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Size measurements, clutch observations, and blood isotope ratios for adult foraging Adélie, Chinstrap, and Gentoo penguins observed on islands in the Palmer Archipelago near Palmer Station, Antarctica. Data were collected and made available by Dr. Kristen Gorman and the Palmer Station Long Term Ecological Research (LTER) Program.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://allisonhorst.github.io/palmerpenguins/}
  \item \url{https://github.com/allisonhorst/palmerpenguins}
  \item Report bugs at \url{https://github.com/allisonhorst/palmerpenguins/issues}
}

}
\author{
\strong{Maintainer}: Allison Horst \email{ahorst@ucsb.edu} (\href{https://orcid.org/0000-0002-6047-5564}{ORCID})

Authors:
\itemize{
  \item Alison Hill \email{apresstats@gmail.com} (\href{https://orcid.org/0000-0002-8082-1890}{ORCID})
  \item Kristen Gorman \email{kbgorman@alaska.edu} (\href{https://orcid.org/0000-0002-0258-9264}{ORCID})
}

}
\keyword{internal}
