% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{path_to_file}
\alias{path_to_file}
\title{Get file path to \code{penguins.csv} and \code{penguins_raw.csv} files}
\source{
This function is adapted from \code{readxl::readxl_example()}.
}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{Name of file in quotes with extension;
\code{"penguins.csv"} and \code{"penguins_raw.csv"} will work.
If \code{NULL}, the example files will be listed.}
}
\description{
palmerpenguins comes bundled with two csv files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\details{
The data in each file is the same as \code{\link{penguins}} and \code{\link{penguins_raw}},
but without the variable types.
}
\examples{
path_to_file()
path_to_file("penguins_raw.csv")
head(read.csv(path_to_file("penguins.csv")))
}
