% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Simplified dataset with observations for 3 penguin species (Adélie, Chinstrap and Gentoo) in Palmer Archipelago, Antarctica}
\format{A tibble with 344 rows and 7 variables:
\describe{
\item{species}{a character string denoting penguin species (Adélie, Chinstrap and Gentoo)}
\item{island}{a character string denoting island in Palmer Archipelago, Antarctica (Biscoe, Dream or Torgersen)}
\item{culmen_length_mm}{a number denoting culmen length (millimeters)}
\item{culmen_depth_mm}{a number denoting culmen depth (millimeters)}
\item{flipper_length_mm}{an integer denoting flipper length (millimeters)}
\item{body_mass_g}{an integer denoting body mass (grams)}
\item{sex}{a character string denoting penguin sex}
}}
\source{
\url{https://doi.org/10.6073/pasta/98b16d7d563f265cb52372c8ca99e60f}

\url{https://doi.org/10.6073/pasta/7fca67fb28d56ee2ffa3d9370ebda689}

\url{https://doi.org/10.6073/pasta/c14dfcfada8ea13a17536e73eb6fbe9e}
}
\usage{
penguins
}
\description{
Includes measurements for penguin species, island in Palmer Archipelago, size (flipper length, body mass, culmen dimensions), and sex.
}
\keyword{datasets}
