% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Size measurements for adult foraging penguins near Palmer Station, Antarctica}
\format{
A tibble with 344 rows and 8 variables:
\describe{
\item{species}{a factor denoting penguin species (Adélie, Chinstrap and Gentoo)}
\item{island}{a factor denoting island in Palmer Archipelago, Antarctica (Biscoe, Dream or Torgersen)}
\item{bill_length_mm}{a number denoting bill length (millimeters)}
\item{bill_depth_mm}{a number denoting bill depth (millimeters)}
\item{flipper_length_mm}{an integer denoting flipper length (millimeters)}
\item{body_mass_g}{an integer denoting body mass (grams)}
\item{sex}{a factor denoting penguin sex (female, male)}
\item{year}{an integer denoting the study year (2007, 2008, or 2009)}
}
}
\source{
{Adélie penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Adélie penguins (Pygoscelis adeliae) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative} \url{https://doi.org/10.6073/pasta/98b16d7d563f265cb52372c8ca99e60f}

{Gentoo penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Gentoo penguin (Pygoscelis papua) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative} \url{https://doi.org/10.6073/pasta/7fca67fb28d56ee2ffa3d9370ebda689}

{Chinstrap penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Chinstrap penguin (Pygoscelis antarcticus) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 6. Environmental Data Initiative} \url{https://doi.org/10.6073/pasta/c14dfcfada8ea13a17536e73eb6fbe9e}

{Originally published in: Gorman KB, Williams TD, Fraser WR (2014) Ecological Sexual Dimorphism and Environmental Variability within a Community of Antarctic Penguins (Genus Pygoscelis). PLoS ONE 9(3): e90081. doi:10.1371/journal.pone.0090081}
}
\usage{
penguins
}
\description{
Includes measurements for penguin species, island in Palmer Archipelago,
size (flipper length, body mass, bill dimensions), and sex.
This is a subset of \code{\link{penguins_raw}}.
}
\keyword{datasets}
