% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{penguins: Palmer Penguins}
\format{
A data frame with 344 rows and 9 variables:
\describe{
\item{rowid}{Unique identifier assigned to each penguin observation. (Range: 1 to 344)}
\item{species}{Species classification of the penguin. (Examples: Adelie, Gentoo, Chinstrap)}
\item{island}{Location of the penguin observation on the island. (Examples: Torgersen, Biscoe, Dream)}
\item{bill_length_mm}{Bill length of the penguin measured in millimeters. (Range: 32.1 to 59.6)}
\item{bill_depth_mm}{Bill depth of the penguin measured in millimeters. (Range: 13.1 to 21.5)}
\item{flipper_length_mm}{Flipper length of the penguin measured in millimeters. (Range: 172 to 231)}
\item{body_mass_g}{Body mass of the penguin measured in grams. (Range: 2700 to 6300)}
\item{sex}{Biological sex of the penguin. (Examples: male, female, NA)}
\item{year}{Year of the penguin observation. (Examples: 2007, 2008, 2009)}
}
}
\usage{
penguins
}
\description{
The Palmer Penguins dataset contains penguins measurements.
}
\keyword{datasets}
